backend/lyrebird/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
bson==0.5.10
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
distro==1.9.0
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.24.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==1.24.4
openai==1.57.2
pika==1.3.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.4
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
faker